@@ -12,20 +12,20 @@ contract('Simple', accounts => {
   it('should set x to 5', () => {
     let simple;
     return Simple.deployed()
-        .then(instance => instance.test(5)) // We need this line to generate some coverage
-        .then(() => {
-          const message = 'Enclosed is my formal application for permanent residency in New Zealand';
-          const messageSha3 = web3.sha3(message);
-          const signature = web3.eth.sign(accounts[0], messageSha3);
+      .then(instance => instance.test(5)) // We need this line to generate some coverage
+      .then(() => {
+        const message = 'Enclosed is my formal application for permanent residency in New Zealand';
+        const messageSha3 = web3.sha3(message);
+        const signature = web3.eth.sign(accounts[0], messageSha3);
 
-          const messageBuffer = new Buffer(messageSha3.replace('0x', ''), 'hex');
-          const messagePersonalHash = ethUtil.hashPersonalMessage(messageBuffer);
+        const messageBuffer = Buffer.from(messageSha3.replace('0x', ''), 'hex');
+        const messagePersonalHash = ethUtil.hashPersonalMessage(messageBuffer);
 
-          const sigParams = ethUtil.fromRpcSig(signature);
-          const publicKey = ethUtil.ecrecover(messagePersonalHash, sigParams.v, sigParams.r, sigParams.s);
-          const senderBuffer = ethUtil.pubToAddress(publicKey);
-          const sender = ethUtil.bufferToHex(senderBuffer);
-          assert.equal(sender, accounts[0]);
-        });
+        const sigParams = ethUtil.fromRpcSig(signature);
+        const publicKey = ethUtil.ecrecover(messagePersonalHash, sigParams.v, sigParams.r, sigParams.s);
+        const senderBuffer = ethUtil.pubToAddress(publicKey);
+        const sender = ethUtil.bufferToHex(senderBuffer);
+        assert.equal(sender, accounts[0]);
+      });
   });
 });
