@@ -11,38 +11,38 @@ const defaultTruffleConfig = require('./truffleConfig.js');
 
 const isWin = /^win/.test(process.platform);
 
-const gasLimitHex = 0xfffffffffff;               // High gas block limit / contract deployment limit
-const gasPriceHex = 0x01;                        // Low gas price
+const gasLimitHex = 0xfffffffffff; // High gas block limit / contract deployment limit
+const gasPriceHex = 0x01; // Low gas price
 
 /**
  * Coverage Runner
  */
 class App {
   constructor(config) {
-    this.coverageDir = './coverageEnv';          // Env that instrumented .sols are tested in
+    this.coverageDir = './coverageEnv'; // Env that instrumented .sols are tested in
 
     // Options
-    this.network = '';                           // Default truffle network execution flag
-    this.silence = '';                           // Default log level passed to shell
+    this.network = ''; // Default truffle network execution flag
+    this.silence = ''; // Default log level passed to shell
     this.log = console.log;
 
     // Other
-    this.testrpcProcess = null;                  // ref to testrpc server we need to close on exit
-    this.events = null;                           // ref to string array loaded from 'allFiredEvents'
-    this.testsErrored = null;                    // flag set to non-null if truffle tests error
-    this.coverage = new CoverageMap();           // initialize a coverage map
+    this.testrpcProcess = null; // ref to testrpc server we need to close on exit
+    this.events = null; // ref to string array loaded from 'allFiredEvents'
+    this.testsErrored = null; // flag set to non-null if truffle tests error
+    this.coverage = new CoverageMap(); // initialize a coverage map
 
     // Config
     this.config = config || {};
-    this.workingDir = config.dir || '.';         // Relative path to contracts folder
-    this.accounts = config.accounts || 35;       // Number of accounts to testrpc launches with
-    this.skipFiles = config.skipFiles || [];     // Which files should be skipped during instrumentation
-    this.norpc = config.norpc || false;          // Launch testrpc-sc internally?
-    this.port = config.port || 8555;             // Port testrpc should listen on
+    this.workingDir = config.dir || '.'; // Relative path to contracts folder
+    this.accounts = config.accounts || 35; // Number of accounts to testrpc launches with
+    this.skipFiles = config.skipFiles || []; // Which files should be skipped during instrumentation
+    this.norpc = config.norpc || false; // Launch testrpc-sc internally?
+    this.port = config.port || 8555; // Port testrpc should listen on
 
-    this.copyNodeModules = config.copyNodeModules || false;  // Copy node modules into coverageEnv?
-    this.testrpcOptions = config.testrpcOptions || null;     // Options for testrpc-sc
-    this.testCommand = config.testCommand || null;           // Optional test command
+    this.copyNodeModules = config.copyNodeModules || false; // Copy node modules into coverageEnv?
+    this.testrpcOptions = config.testrpcOptions || null; // Options for testrpc-sc
+    this.testCommand = config.testCommand || null; // Optional test command
 
     this.setLoggingLevel(config.silent);
   }
@@ -75,8 +75,8 @@ class App {
       let truffleConfig;
 
       shell.test('-e', `${this.workingDir}/truffle.js`)
-          ? truffleConfig = reqCwd.silent(`${this.workingDir}/truffle.js`)
-          : truffleConfig = reqCwd.silent(`${this.workingDir}/truffle-config.js`);
+        ? truffleConfig = reqCwd.silent(`${this.workingDir}/truffle.js`)
+        : truffleConfig = reqCwd.silent(`${this.workingDir}/truffle-config.js`);
 
       // Coverage network opts specified: use port if declared
       if (truffleConfig && truffleConfig.networks && truffleConfig.networks.coverage) {
@@ -91,7 +91,6 @@ class App {
           ? fs.writeFileSync(`${this.coverageDir}/truffle-config.js`, trufflejs)
           : fs.writeFileSync(`${this.coverageDir}/truffle.js`, trufflejs);
       }
-
     } catch (err) {
       const msg = ('There was a problem generating the coverage environment: ');
       this.cleanUp(msg + err);
@@ -117,7 +116,7 @@ class App {
           this.log('Instrumenting ', file);
 
           currentFile = file;
-          const contractPath = this.platformNeutralPath(file);
+          const contractPath = App.platformNeutralPath(file);
           const working = this.workingDir.substring(1);
           const canonicalPath = contractPath.split('/coverageEnv').join(working);
           const contract = fs.readFileSync(contractPath).toString();
@@ -132,7 +131,7 @@ class App {
       const msg = `There was a problem instrumenting ${currentFile}: `;
       this.cleanUp(msg + err);
     }
-    this.postProcessPure(this.coverageDir);
+    App.postProcessPure(this.coverageDir);
   }
 
   /**
@@ -148,7 +147,9 @@ class App {
         const command = './node_modules/ethereumjs-testrpc-sc/build/cli.node.js ';
 
         // Launch
-        const execOpts = {maxBuffer: 1024 * 1024 * 10};
+        const execOpts = {
+          maxBuffer: 1024 * 1024 * 10,
+        };
         this.testrpcProcess = childprocess.exec(command + options, execOpts, (err, stdout, stderr) => {
           if (err) {
             if (stdout) this.log(`testRpc stdout:\n${stdout}`);
@@ -206,7 +207,7 @@ class App {
 
     return new Promise((resolve, reject) => {
       // Get events fired during instrumented contracts execution.
-      const stream = fs.createReadStream(`./allFiredEvents`);
+      const stream = fs.createReadStream('./allFiredEvents');
       stream.on('error', err => this.cleanUp('Event trace could not be read.\n' + err));
       const reader = readline.createInterface({
         input: stream,
@@ -218,7 +219,7 @@ class App {
           // Generate Istanbul report
           try {
             this.coverage.generate(this.events, `${this.workingDir}/contracts`);
-            const relativeMapping = this.makeKeysRelative(this.coverage.coverage, this.workingDir);
+            const relativeMapping = this.makeKeysRelative(this.workingDir);
             const json = JSON.stringify(relativeMapping);
             fs.writeFileSync('./coverage.json', json);
 
@@ -242,12 +243,12 @@ class App {
 
   // ------------------------------------------ Utils ----------------------------------------------
 
-  sanityCheckContext(){
-    if (!shell.test('-e', `${this.workingDir}/contracts`)){
-      this.cleanUp("Couldn't find a 'contracts' folder to instrument.");
+  sanityCheckContext() {
+    if (!shell.test('-e', `${this.workingDir}/contracts`)) {
+      this.cleanUp('Couldn\'t find a \'contracts\' folder to instrument.');
     }
 
-    if (shell.test('-e', `${this.workingDir}/${this.coverageDir}`)){
+    if (shell.test('-e', `${this.workingDir}/${this.coverageDir}`)) {
       shell.rm('-Rf', this.coverageDir);
     }
   }
@@ -258,10 +259,17 @@ class App {
    * @param  {[type]} root working directory
    * @return {[type]}      map with relativized keys
    */
-  makeKeysRelative(map, root) {
+  makeKeysRelative(root) {
     const newCoverage = {};
-    Object.keys(map).forEach(pathKey => {
-      newCoverage[path.relative(root, pathKey)] = map[pathKey];
+    // NOTE: just a personal preference but it would be much cleaner for linting to
+    //       allow
+    //       const { coverage } = this.coverage;
+    // ref rule: object-curly-newline
+    const {
+      coverage,
+    } = this.coverage;
+    Object.keys(coverage).forEach(pathKey => {
+      newCoverage[path.relative(root, pathKey)] = coverage[pathKey];
     });
     return newCoverage;
   }
@@ -271,7 +279,7 @@ class App {
    * @param  {String} file path
    * @return {[type]}      normalized path
    */
-  platformNeutralPath(file) {
+  static platformNeutralPath(file) {
     return (isWin)
       ? path.resolve(file).split('\\').join('/')
       : path.resolve(file);
@@ -282,18 +290,18 @@ class App {
    * in the coverageEnv before the `contracts` folder is instrumented.
    * @param  {String} env 'coverageEnv' presumably
    */
-  postProcessPure(env) {
+  static postProcessPure(env) {
     shell.ls(`${env}/**/*.sol`).forEach(file => {
       const pureRe = /\spure\s/gi;
       const viewRe = /\sview\s/gi;
       const constantRe = /\sconstant\s/gi;
-      const contractPath = this.platformNeutralPath(file);
+      const contractPath = App.platformNeutralPath(file);
       let contract = fs.readFileSync(contractPath).toString();
       contract = contract.replace(pureRe, ' ');
       contract = contract.replace(viewRe, ' ');
       contract = contract.replace(constantRe, ' ');
       fs.writeFileSync(contractPath, contract);
-    })
+    });
   }
 
   /**
