@@ -1,4 +1,4 @@
-const SolidityParser = require('solidity-parser-sc');
+const SolidityParser = require('solparse');
 const preprocessor = require('./preprocessor');
 const injector = require('./injector');
 const parse = require('./parse');
