@@ -1,5 +1,5 @@
 const SolExplore = require('sol-explore');
-const SolidityParser = require('solidity-parser-sc');
+const SolidityParser = require('solparse');
 
 /**
  * Splices enclosing brackets into `contract` around `expression`;
@@ -37,7 +37,7 @@ module.exports.run = function r(contract) {
           this.stopTraversal();
         // If alternates
         } else if (
-            node.type === 'IfStatement' &&
+          node.type === 'IfStatement' &&
             node.alternate &&
             node.alternate.type !== 'BlockStatement') {
           contract = blockWrap(contract, node.alternate);
@@ -45,7 +45,7 @@ module.exports.run = function r(contract) {
           this.stopTraversal();
         // Loops
         } else if (
-            (node.type === 'ForStatement' || node.type === 'WhileStatement') &&
+          (node.type === 'ForStatement' || node.type === 'WhileStatement') &&
             node.body.type !== 'BlockStatement') {
           contract = blockWrap(contract, node.body);
           keepRunning = true;
