@@ -6,9 +6,9 @@ const Proxy = artifacts.require('./Proxy.sol');
 
 contract('Proxy', accounts => {
   it('Should compile and run when one contract inherits from another', () => Owned.deployed()
-        .then(() => Proxy.deployed())
-        .then(instance => instance.isOwner.call({
-          from: accounts[0],
-        }))
-        .then(val => assert.equal(val, true)));
+    .then(() => Proxy.deployed())
+    .then(instance => instance.isOwner.call({
+      from: accounts[0],
+    }))
+    .then(val => assert.equal(val, true)));
 });
