@@ -11,7 +11,7 @@ contract('SimpleError', () => {
       simple = instance;
       return simple.test(5);
     })
-    .then(() => simple.getX.call())
-    .then(val => assert.equal(val, 5));
+      .then(() => simple.getX.call())
+      .then(val => assert.equal(val, 5));
   });
 });
