@@ -10,7 +10,7 @@ contract('Simple', () => {
       simple = instance;
       return simple.test(5);
     })
-    .then(() => simple.getX.call())
-    .then(val => assert.equal(val.toNumber(), 4)); // <-- Wrong result: test fails
+      .then(() => simple.getX.call())
+      .then(val => assert.equal(val.toNumber(), 4)); // <-- Wrong result: test fails
   });
 });
